Adding initsystem fact & required java version
To remedy problems with processing of required java version both in
production code and in tests I've added required_java function.

The required_java function can calculate required version of Java
to be used with given version of JBoss or Wildfly.

The second part is jboss_initsystem fact. This fact produce what
initsystem is the system running on: SystemV or SystemD.