Replacing SystemD and Sys V on Beaker on Docker
Using System V and SystemD init systems is not ideal. They tend to
interfere heavily with host's own init systems - processes are going
bad, uses 100% CPU, or just crush. That's because thay are run in
privileged mode.

This commit tries to fix that be switching over to replacements of those
init systems. For System V, that's custom supervisord running SSH, and
for SystemD that's gdraheim/docker-systemctl-replacement python script.

GitHub issue: #104